requirements.txt
blinker==1.9.0
cachelib==0.13.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.0.0
Flask-Session==0.5.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
mysql-connector-python==8.2.0
packaging==25.0
pillow==12.0.0
protobuf==4.21.12
python-dotenv==1.0.0
reportlab==4.4.5
Werkzeug==3.0.1
